Add LIBERO/LIBERO-plus training configs, marigold DROID data support, and eval scripts
Also stop tracking starVLA_working (accidentally added as a nested repo) and
gitignore it going forward.

+++ starVLA/model/modules/vlm/CosmosReason2.py
@@ -57,7 +57,7 @@ class _CosmosReason2_Interface(nn.Module):
         self,
         **kwargs,
     ):
-        with torch.autocast("cuda", dtype=torch.float16):
+        with torch.autocast("cuda", dtype=torch.bfloat16):
             generation_output = self.model.generate(
                 **kwargs,
             )
