+++ starVLA/model/modules/vlm/Molmo2.py
@@ -117,7 +117,7 @@ class _Molmo2_VL_Interface(nn.Module):
         return outputs
 
     def generate(self, **kwargs):
-        with torch.autocast("cuda", dtype=torch.float16):
+        with torch.autocast("cuda", dtype=torch.bfloat16):
             return self.model.generate(**kwargs)
 
     def build_qwenvl_inputs(self, images, instructions, solutions=None, **kwargs):
